refactor: abstract product card into another component
